requirements.txt
# PyTorch and related packages
torch==2.7.1+cu118
torchvision==0.22.1+cu118
torchaudio==2.7.1+cu118

# Deep Learning Models
segmentation-models-pytorch>=0.5.0
timm>=1.0.0

# Image Processing
imgaug==0.4.0
opencv-python>=4.11.0
SimpleITK>=2.5.0
scikit-image>=0.25.0

# Data Processing
pandas>=2.2.0
numpy>=1.26.0
scikit-learn>=1.5.0

# Visualization
matplotlib>=3.9.0
visdom>=0.2.4
tensorboard>=2.20.0
tensorboardX>=2.6.0

# Hyperparameter Tuning
ray[tune]>=2.50.0

# Utilities
PyYAML>=6.0.0
tqdm>=4.66.0
pyarrow>=16.0.0

